06-10

+++ marco/minishell_tmp/env.c
@@ -6,7 +6,7 @@
 /*   By: acosta-a <acosta-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/17 00:57:01 by acosta-a          #+#    #+#             */
-/*   Updated: 2022/10/01 08:16:31 by acosta-a         ###   ########.fr       */
+/*   Updated: 2022/10/05 17:00:17 by acosta-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,10 @@
 int	ft_env(t_data **data, char *input)
 {
 	int	i;
+	int	p;
 
 	i = 0;
+	p = getpid();
 	if (input && input[0] != '\0' && input[0] != '|' && input[0]
 		!= '>' && input[0] != ';')
 	{
@@ -28,8 +30,7 @@ int	ft_env(t_data **data, char *input)
 		ft_putstrs((*data)->envp[i], "\n", 0, 1);
 		i++;
 	}
-	int p = getpid();
-	kill(p, SIGTERM);
+	if (p > 0)
+		kill(p, SIGTERM);
 	return (0);
-
 }
